+++ lib/assets/javascripts/new-dashboard/components/MapCard/SimpleMapCard.vue
@@ -9,14 +9,15 @@
        'card--can-hover': canHover
      }"
     @click="onClick">
-    <div class="card-media" :class="{'has-error': isThumbnailErrored}">
-      <img :src="mapThumbnailUrl" @error="onThumbnailError" v-if="isMap && !isThumbnailErrored"/>
+    <div class="card-media" :class="{ 'is-kuviz': isKuviz ,'has-error': !isKuviz && isThumbnailErrored  }">
+      <img :src="mapThumbnailUrl" @error="onThumbnailError" v-if="isBuilderMap && !isThumbnailErrored"/>
 
-      <div class="media-dataset" v-if="!isMap">
+      <div class="media-dataset" v-if="!(isBuilderMap || isKuviz)">
         <img svg-inline src="../../assets/icons/datasets/dataset-icon.svg" />
       </div>
 
-      <div class="MapCard-error" v-if="isThumbnailErrored"></div>
+      <span class="card-type title is-xsmall is-semibold">{{ $t(`MapCard.type.${visualization.type}`) }}</span>
+      <div class="MapCard-error" v-if="!isKuviz && isThumbnailErrored"></div>
     </div>
 
     <span class="checkbox card-select" @mouseover="mouseOverChildElement" @mouseleave="mouseOutChildElement">
@@ -136,7 +137,7 @@ export default {
         return 'DatasetQuickActions';
       }
 
-      if (visualizationType === 'derived') {
+      if (visualizationType === 'derived' || visualizationType === 'kuviz') {
         return 'MapQuickActions';
       }
     },
@@ -149,8 +150,11 @@ export default {
         return allSections;
       }, {});
     },
-    isMap () {
+    isBuilderMap () {
       return this.$props.visualization.type === 'derived';
+    },
+    isKuviz () {
+      return this.$props.visualization.type === 'kuviz';
     }
   },
   methods: {
@@ -261,6 +265,17 @@ export default {
   color: $text__color;
 }
 
+.card-type {
+  position: absolute;
+  bottom: 12px;
+  left: 12px;
+  padding: 6px 8px;
+  border-radius: 2px;
+  border-color: $card__border-color;
+  background: $white;
+  text-transform: uppercase;
+}
+
 .card-media {
   display: flex;
   position: relative;
@@ -268,6 +283,15 @@ export default {
   overflow: hidden;
   background: url($assetsDir + '/images/layout/default-map-bkg.png') no-repeat center 0;
   background-size: cover;
+
+  &.is-kuviz {
+    display: block;
+    background: url($assetsDir + '/images/layout/kuviz-map-bkg.png');
+
+    .card-type {
+      color: $card-cf__color;
+    }
+  }
 
   img {
     width: 100%;